requirements.txt
quart
flask-login
tortoise-orm
hypercorn
pydantic[dotenv]
